rename guild tests
